Formatting Text
Fixed Issues with Subheadings in which the sub headings wrap around by one letter and over all looks very nasty for a website page to have. especailly when there is room to make the sentance one line. Fixed Spacing on News page in which News was more indented then it's subheading making it look unorangized. I'm referencing how a Online News artilce might look for formatting.