Minor Fixes from last commit : 6/03
The fixes invole word wraping when it shouldn't be. Like publications was wrapping before but shouldn't anymore. The reason for this was wix.com used a "break-word" on overflow and word wrap so it would break the word instead of the sentance. Fixed Social Media bar at bottom of page.