Ethics Update
Turns out the footpprints go overscreen and cause an issue in that before this patch they would be a layer above the social icons meaning you couldn't click them. That's fixed and when the footprints went off screen they would extend the scrollable off screen but I fixed that by turining the overflow of the footprints div to hidden.