Sidebar and Copyright Inclusion and Changes
Included images in the placeholder <img> in the side bar and included the copyrights for my images and other images. now Secna park has copyright on everything and you need to ask for permission but the terms state any user can take pictures within the park and outside with vistors. So that's how I got the photo, which a photography gave to wikipedia under CC.